Refactored plugin Changed Overlay and version
Added Wrath rune support.
Added Farming cape support for banking/home tele if the cape is equipped for Blood Runes. - WrathSupport for farming cape in future.

other minor changes in various areas
