+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/frosty/frostyrc/RcOverlay.java
@@ -1,4 +1,4 @@
-package net.runelite.client.plugins.microbot.frosty.bloods;
+package net.runelite.client.plugins.microbot.frosty.frostyrc;
 
 import net.runelite.client.game.ItemManager;
 import net.runelite.client.plugins.microbot.Microbot;
@@ -15,12 +15,12 @@ import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Graphics2D;
 
-public class BloodsOverlay extends OverlayPanel {
-    private final BloodsPlugin plugin;
+public class RcOverlay extends OverlayPanel {
+    private final RcPlugin plugin;
     private final DateTimeFormatter dtf = DateTimeFormatter.ofPattern("HH:mm:ss");
 
     @Inject
-    BloodsOverlay(BloodsPlugin plugin) {
+    RcOverlay(RcPlugin plugin) {
         super(plugin);
         this.plugin = plugin;
         setPosition(OverlayPosition.TOP_LEFT);
@@ -40,7 +40,7 @@ public class BloodsOverlay extends OverlayPanel {
             //panelComponent.getChildren().add(new ImageComponent(bloodRuneImage));
 
             panelComponent.getChildren().add(TitleComponent.builder()
-                    .text("❄️ True Blood: " + BloodsPlugin.getVersion()).color((Color.red)).build());
+                    .text("❄️ Frosty Rc " + RcPlugin.getVersion()).color((Color.cyan)).build());
 
             panelComponent.getChildren().add(LineComponent.builder().left("Xp gained:")
                     .right(String.valueOf(plugin.getTotalXpGained())).build());
